backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
google-generativeai==0.8.3
numpy==1.26.4
pydantic==2.10.3
python-dotenv==1.0.1